requirements.txt
Flask==0.10.1
Flask-Fanstatic==0.2.0
Flask-Login==0.2.9
Flask-Mail==0.9.0
Flask-Principal==0.4.0
Flask-RESTful==0.2.11
Flask-Script==0.6.6
Flask-Security==1.7.1
Flask-SuperAdmin==1.7.1
Flask-WTF==0.9.4
Jinja2==2.7.2
MarkupSafe==0.18
WTForms==1.0.5
WebOb==1.3.1
Werkzeug==0.9.4
aniso8601==0.82
argparse==1.2.1
blinker==1.3
fanstatic==0.16
flask-mongoengine==0.7.0
itsdangerous==0.23
mongoengine==0.8.6
passlib==1.6.2
pycrypto==2.6.1
pymongo==2.6.3
pytz==2013.9
six==1.5.2
wsgiref==0.1.2
